Fix screen scale issue. Correct images now loading

+++ IGraphics/Platforms/IGraphicsWin.cpp
@@ -1172,7 +1172,7 @@ void IGraphicsWin::UpdateTooltips()
 void* IGraphicsWin::OpenWindow(void* pParent)
 {
   mParentWnd = (HWND) pParent;
-  int screenScale = GetScaleForHWND(mParentWnd);
+  float screenScale = GetScaleForHWND(mParentWnd);
   int x = 0, y = 0, w = WindowWidth() * screenScale, h = WindowHeight() * screenScale;
 
   if (mPlugWnd)
